requirements.txt
asyncpg==0.30.0
fastapi==0.115.14
psycopg2-binary==2.9.10
python-dotenv==1.1.1
SQLAlchemy==2.0.41
uvicorn==0.34.3
Faker==37.4.0
geopy==2.4.1
